Fixes typo in name
