+++ LibDmd/Input/PinballFX/PinballFX3MemoryGrabberBase.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Diagnostics;
+
+namespace LibDmd.Input.PinballFX
+{
+	public abstract class PinballFX3MemoryGrabberBase<FrameType>: MemoryGrabber<FrameType>
+	{
+		protected override IntPtr AttachGameProcess(Process p)
+		{
+			if (p.ProcessName == "Pinball FX3")
+			{
+				return GetPointerBaseAddress(p);
+			}
+			return IntPtr.Zero;
+		}
+
+		abstract protected IntPtr GetPointerBaseAddress(Process gameProc);
+	}
+}
